Moved the service registration to async_setup_entry as this seems to be the more common approach.
