+++ homeassistant/components/yamaha_musiccast/__init__.py
@@ -9,8 +9,7 @@ from aiomusiccast.musiccast_device import MusicCastData, MusicCastDevice
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST
-from homeassistant.core import HomeAssistant, ServiceCall
-from homeassistant.helpers import service
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC, format_mac
 from homeassistant.helpers.entity import DeviceInfo
@@ -20,75 +19,12 @@ from homeassistant.helpers.update_coordinator import (
     UpdateFailed,
 )
 
-from .const import (
-    ATTR_MASTER,
-    BRAND,
-    DOMAIN,
-    JOIN_SERVICE_SCHEMA,
-    SERVICE_JOIN,
-    SERVICE_UNJOIN,
-    UNJOIN_SERVICE_SCHEMA,
-)
+from .const import BRAND, DOMAIN
 
 PLATFORMS = ["media_player"]
 
 _LOGGER = logging.getLogger(__name__)
 SCAN_INTERVAL = timedelta(seconds=60)
-
-
-async def async_setup(hass: HomeAssistant, config: dict):
-    """Set up the MusicCast component."""
-
-    @service.verify_domain_control(hass, DOMAIN)
-    async def async_group_service_handle(service_call: ServiceCall):
-        """Handle services."""
-        entity_ids = service_call.data.get("entity_id", [])
-        if not entity_ids:
-            return
-
-        all_entities = []
-        for coord in hass.data[DOMAIN].values():
-            all_entities += coord.entities
-
-        entities = [entity for entity in all_entities if entity.entity_id in entity_ids]
-
-        if service_call.service == SERVICE_JOIN:
-            master_id = service_call.data[ATTR_MASTER]
-            master = next(
-                (entity for entity in all_entities if entity.entity_id == master_id),
-                None,
-            )
-            if master and isinstance(master, MusicCastDeviceEntity):
-                await master.async_server_join(entities)
-            else:
-                _LOGGER.error(
-                    "Invalid master specified for join service: %s",
-                    service_call.data[ATTR_MASTER],
-                )
-        elif service_call.service == SERVICE_UNJOIN:
-            for entity in entities:
-                if isinstance(entity, MusicCastDeviceEntity):
-                    await entity.async_unjoin()
-                else:
-                    _LOGGER.error(
-                        "Invalid entity specified for unjoin service: %s",
-                        entity,
-                    )
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_JOIN,
-        async_group_service_handle,
-        JOIN_SERVICE_SCHEMA,
-    )
-
-    hass.services.async_register(
-        DOMAIN,
-        SERVICE_UNJOIN,
-        async_group_service_handle,
-        UNJOIN_SERVICE_SCHEMA,
-    )
-    return True
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
